# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/listings": {
            "get": {
                "description": "Get details of all listings",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "listings"
                ],
                "summary": "Get details of all listings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token header",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Listing"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new listing",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "listings"
                ],
                "summary": "Create listings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token header",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Listing"
                            }
                        }
                    }
                }
            }
        },
        "/listings/{listingId}": {
            "get": {
                "description": "Get details of the listing",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "listings"
                ],
                "summary": "Get details of listing",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token header",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "listingId",
                        "name": "leaseId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Listing"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update the existing listing",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "listings"
                ],
                "summary": "Update listing",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token header",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Listing"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete the existing lease",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "listings"
                ],
                "summary": "Delete listings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token header",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Listing"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "models.Listing": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "houseType": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isleased": {
                    "type": "boolean"
                },
                "listingImg": {
                    "type": "string"
                },
                "listingType": {
                    "type": "string"
                },
                "rent": {
                    "type": "integer"
                },
                "societyId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
...
    }
}